IntMath test for opcode parsing and execution
Unfortunately, had to use static fields in order to keep the math
operations; otherwise the compiler did constant expression elimination;
it just loaded and printed the results :( To really make tests we'll
need to create a bunch of methods, each of which will trigger an
appropriate opcode. This will come later.